+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - reverse / flip the matrix method added
 - `comparseobject` deprecated in favor of `compare`
+- `concatenate`, `diagonal`, `reshape`, `transpose` and `flatten` are vectorized now
 
 ## 1.0.3